HW13: refactor SqlStorage need use DRY AllStorageTest is passed

+++ src/ru/GilvanovDr/WebApp/storage/SqlStorage.java
@@ -28,8 +28,8 @@ public class SqlStorage implements Storage {
              PreparedStatement ps = con.prepareStatement("SELECT * FROM resume")) {
             ResultSet rs = ps.executeQuery();
             List<Resume> list = new ArrayList<>();
-            while (!rs.next()) {
-                list.add(new Resume(rs.getString("uuid"), rs.getString("full_name")));
+            while (rs.next()) {
+                list.add(new Resume(rs.getString("uuid").trim(), rs.getString("full_name")));
             }
             return list;
         } catch (SQLException e) {
@@ -40,16 +40,17 @@ public class SqlStorage implements Storage {
     // todo  Use SqlException code! remove doDelete, doSave
     @Override
     public void update(Resume resume) {
-        try (Connection con = connectionFactory.getConnection();
-             PreparedStatement ps = con.prepareStatement("UPDATE resume SET full_name=? WHERE uuid=?")) {
-            ps.setString(1, resume.getFullName());
-            ps.setString(2, resume.getUuid());
-            ps.execute();
-        } catch (SQLException e) {
-            if ("23505".equals(e.getSQLState())) {
-                throw new ExistStorageException(resume.getUuid());
+        if (!isExist(resume.getUuid())) {
+            throw new NoExistStorageException(resume.getUuid());
+        } else {
+            try (Connection con = connectionFactory.getConnection();
+                 PreparedStatement ps = con.prepareStatement("UPDATE resume SET full_name=? WHERE uuid=?")) {
+                ps.setString(1, resume.getFullName());
+                ps.setString(2, resume.getUuid());
+                ps.execute();
+            } catch (SQLException e) {
+                throw new StorageException(e);
             }
-            throw new StorageException(e);
         }
     }
 
@@ -57,8 +58,8 @@ public class SqlStorage implements Storage {
     public void save(Resume r) {
         try (Connection con = connectionFactory.getConnection();
              PreparedStatement ps = con.prepareStatement("INSERT INTO resume (uuid, full_name) VALUES (?,?)")) {
-            ps.setString(1,r.getUuid());
-            ps.setString(2,r.getFullName());
+            ps.setString(1, r.getUuid());
+            ps.setString(2, r.getFullName());
             ps.execute();
         } catch (SQLException e) {
             if ("23505".equals(e.getSQLState())) {
@@ -68,8 +69,7 @@ public class SqlStorage implements Storage {
         }
     }
 
-    @Override
-    public void delete(String uuid) {
+    private boolean isExist(String uuid) {
         try (Connection con = connectionFactory.getConnection();
              PreparedStatement ps = con.prepareStatement("SELECT EXISTS(SELECT uuid FROM resume WHERE uuid = ?)")) {
             ps.setString(1, uuid);
@@ -77,24 +77,26 @@ public class SqlStorage implements Storage {
             if (!rs.next()) {
                 throw new StorageException("Can't read from db", null);
             } else {
-                if (!rs.getBoolean("exists")) {
-                    throw new NoExistStorageException(uuid);
-                } else {
-                    doDelete(uuid);
-                }
+                return rs.getBoolean("exists");
             }
         } catch (SQLException e) {
             throw new StorageException(e);
         }
     }
 
-    private void doDelete(String uuid) {
-        try (Connection con = connectionFactory.getConnection();
-             PreparedStatement ps = con.prepareStatement("DELETE FROM resume WHERE uuid=?")) {
-            ps.setString(1, uuid);
-            ps.execute();
-        } catch (SQLException e) {
-            throw new StorageException(e);
+    @Override
+    public void delete(String uuid) {
+        if (!isExist(uuid)) {
+            throw new NoExistStorageException(uuid);
+        } else {
+            try (Connection con = connectionFactory.getConnection();
+                 PreparedStatement ps = con.prepareStatement("DELETE FROM resume WHERE uuid=?")) {
+                ps.setString(1, uuid);
+                ps.execute();
+
+            } catch (SQLException e) {
+                throw new StorageException(e);
+            }
         }
     }
 
@@ -118,7 +120,6 @@ public class SqlStorage implements Storage {
         try (Connection con = connectionFactory.getConnection();
              PreparedStatement ps = con.prepareStatement("SELECT * FROM resume r WHERE r.uuid=?")) {
             ps.setString(1, uuid);
-            ps.execute();
             ResultSet rs = ps.executeQuery();
             if (!rs.next()) {
                 throw new NoExistStorageException(uuid);
